requirement.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
get==2019.4.13
idna==2.8
mysql-connector==2.2.9
mysql-connector-python==8.0.16
post==2019.4.13
protobuf==3.8.0
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
